Crate reth_beacon_consensus

Beacon consensus implementation.

Modules

hooks

Hooks for running during the main loop of consensus engine.

Structs

BeaconConsensus

Ethereum beacon consensus

BeaconConsensusEngine

The beacon consensus engine is the driver that switches between historical and live sync.

BeaconConsensusEngineHandle

A shareable beacon consensus frontend type. Used to interact with the spawned beacon consensus engine task.

Enums

BeaconConsensusEngineError

The error type for the beacon consensus engine service BeaconConsensusEngine

BeaconConsensusEngineEvent

Events emitted by crate::BeaconConsensusEngine.

BeaconEngineMessage

A message for the beacon engine from other components of the node (engine RPC API invoked by the consensus layer).

BeaconForkChoiceUpdateError

Represents error cases for an applied forkchoice update.

BeaconOnNewPayloadError

Represents all error cases when handling a new payload.

ForkchoiceStatus

A simplified representation of PayloadStatusEnum specifically for FCU.

Constants

MIN_BLOCKS_FOR_PIPELINE_RUN

The largest gap for which the tree will be used for sync. See docs for pipeline_run_threshold for more information.

Type Aliases

BeaconEngineResult

Beacon engine result.