Module reth_beacon_consensus::hooks

Hooks for running during the main loop of consensus engine.

Structs

EngineContext

Engine context passed to the hook polling function.

EngineHooks

Collection of engine hooks.

PruneHook

Manages pruning under the control of the engine.

SnapshotHook

Manages snapshotting under the control of the engine.

Enums

EngineHookDBAccessLevel

Level of database access the hook needs for execution.

EngineHookError

An error returned by hook.

EngineHookEvent

An event emitted when hook is polled.

Traits

EngineHook

Hook that will be run during the main loop of consensus engine.